Split external automation manager
